common: slot_vector moved into common
