+++ src/common/slot_vector.h
